refactor: use of placeholders for logging
This changes the way parameters are passed to form logging messages to
use placeholders `{}`. By using placeholders the arguments to logging
framework are not converted to String via (`toString`) and concatenated.
This decreases memory and processing cost by not accumulating memory in
String concatenation and invoking `toString` -- which in case of Jackson
, on ObjectNode classes leads to serialization to JSON.